Make the gallery the homepage
