+++ pages/index.tsx
@@ -1,48 +1,24 @@
 import Head from 'next/head'
 import Layout, { siteTitle } from '../components/layout'
-import utilStyles from '../styles/utils.module.css'
-import { getSortedPostsData } from '../lib/posts'
-import Link from 'next/link'
-import Date from '../components/date'
+import Gallery from "react-photo-gallery"
+import { getAllPhotos } from '../lib/photos'
 
-export default function Home({ allPostsData }) {
+export default function Home({ allPhotosData }) {
   return (
     <Layout home>
       <Head>
         <title>{siteTitle}</title>
       </Head>
-      <section className={utilStyles.headingMd}>
-        <p>Still learning</p>
-        <p>
-          (This is a sample website - you’ll be building a site like this on{' '}
-          <a href="https://nextjs.org/learn">our Next.js tutorial</a>.)
-        </p>
-      </section>
-      <section className={`${utilStyles.headingMd} ${utilStyles.padding1px}`}>
-        <h2 className={utilStyles.headingLg}>Blog</h2>
-        <ul className={utilStyles.list}>
-          {allPostsData.map(({ id, date, title }) => (
-            <li className={utilStyles.listItem} key={id}>
-              <Link href="/posts/[id]" as={`/posts/${id}`}>
-                <a>{title}</a>
-              </Link>
-              <br />
-              <small className={utilStyles.lightText}>
-                <Date dateString={date} />
-              </small>
-            </li>
-          ))}
-        </ul>
-      </section>
+      <Gallery photos={allPhotosData} />
     </Layout>
   )
 }
 
 export async function getStaticProps() {
-  const allPostsData = getSortedPostsData()
+  const allPhotosData = getAllPhotos()
   return {
     props: {
-      allPostsData
+      allPhotosData: allPhotosData
     }
   }
 }
